mlem

+++ VroomInsurance/VroomInsurance/ViewModels/ApplicationVM.cs
@@ -7,7 +7,7 @@ namespace VroomInsurance.ViewModels
 {
     public class ApplicationVM
     {
-        public float Quote { get; set; }
+        public decimal Quote { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string EmailAddress { get; set; }
